Fix test

+++ LibraryCPP/Tests/vector.cpp
@@ -17,7 +17,7 @@ int main()
 
     for (size_t i = 0 ; i < vector_size(vector) ; ++i)
     {
-        if (vector_get(vector, i) != i)
+        if (vector_get(vector, i) != (int)i)
         {
             std::cout << "Invalid vector element " << i << "\n";
             return 1;
@@ -45,7 +45,7 @@ int main()
 
     for (size_t i = 0 ; i < vector_size(vector) ; ++i)
     {
-        if (vector_get(vector, i) != i)
+        if (vector_get(vector, i) != (int)i)
         {
             std::cout << "Invalid vector element " << i << "\n";
             return 1;
